Extracting out the Dropdowns to a separate component, renamed a bunch of component stuff and extracted out elements and title. The updating the firestore document will need revisiting!
